Update App.jsx, TabButton.jsx
conditional and dynamic styling; added a new boolean prop to the button component. If a button is clicked, set the boolean to true, and if the boolean is set to true then set the classname of the button to "active", which have different styling properties in the css file than a button with no classname.